Moved validator plugin to validator package and cleaned up validators init
